Add the UI for the game using Html DOM
In our UI, the player should be able to play the game by clicking on buttons rather than typing their answer in a prompt. Create three buttons, one for each selection. Add an event listener to the buttons that call your playRound function with the correct playerSelection every time a button is clicked. Add a div for displaying results and change all of your console.logs into DOM methods. Display the running score, and announce a winner of the game once one player reaches 5 points. Change if statements to switch statements when possible. Create more functions to make the running of the game smoother.